[AT] update: Change scss to fit teletubbies convention
