+++ components/AlocholType/ATCard/ATCard.module.scss
@@ -1,26 +1,47 @@
 @import '~styles/theme.scss';
 
 .card {
+  width: auto;
+  height: 30vh;
+  border-radius: 10px;
+  padding: 0.5rem;
+
   display: flex;
   flex-direction: row;
   gap: 5%;
-  padding: 0.5rem;
-  width: auto;
-  height: 30vh;
+  @media screen and (min-width: $breakpoint-lg) {
+    // desktop
+    width: 100%;
+    max-height: 40vh;
+    overflow-wrap: break-word;
+  }
+  @media screen and (min-width: $breakpoint-sm) and (max-width: $breakpoint-lg) {
+    // tablet
+    height: 25vh;
+  }
+  @media screen and (max-width: $breakpoint-sm) {
+    //mobile
+    height: auto;
+    flex-direction: column;
+  }
+  &:hover {
+    background-color: rgba($gray11, 0.2);
+    transition: all 0.4s ease-in-out;
+  }
 }
 .card:hover {
   box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.1);
 }
-.alcoholImageBox {
-  width: 25%;
-  aspect-ratio: 5/8;
-  border-radius: 10px;
-  box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.1);
-}
 .alcoholImage {
-  width: 100%;
-  height: 100%;
+  width: 200px;
+  aspect-ratio: 1.618;
   object-fit: cover;
+  border-radius: 10px;
+  @media screen and (max-width: $breakpoint-sm) {
+    width: 100%;
+    aspect-ratio: 2.618;
+    object-fit: scale-down;
+  }
 }
 .alcoholInfoBox {
   width: auto;
@@ -28,36 +49,61 @@
   flex-direction: column;
   justify-content: flex-start;
   align-items: left;
-  gap: 0.3rem;
-  font-size: 0.9rem;
+  gap: 0.5rem;
   overflow-x: hidden;
+  @media screen and (max-width: $breakpoint-sm) {
+    gap: 0.7rem;
+  }
 }
 
 .title {
+  max-width: 100%;
   margin: 0;
-  font-size: $font-size-xxl;
+
+  font-size: $font-size-xl;
   font-weight: bold;
-  color: #000;
+
+  @include ellipsisText(2, $font-size-xl);
+
+  @media screen and (max-width: $breakpoint-sm) {
+    margin-top: 0.7rem;
+    font-size: $font-size-xl;
+    @include ellipsisText(1, $font-size-xl);
+  }
 }
 .basic {
   margin: 0;
-  font-size: 1rem;
+
+  font-size: $font-size-sm;
+  @include ellipsisText(1, $font-size-md);
+
+  @media screen and (max-width: $breakpoint-sm) {
+    font-size: $font-size-sm;
+    @include ellipsisText(1, $font-size-sm);
+  }
 }
 .hashtagBox {
+  max-height: calc((2 * $font-size-xxs) + 30px);
+
   display: flex;
   flex-direction: row;
   justify-content: flex-start;
   align-items: center;
   flex-wrap: wrap;
+  gap: 10px;
 
   padding-top: 0.2rem;
 
-  gap: 10px;
-  font-size: 0.75rem;
+  font-size: $font-size-xxs;
+  overflow: hidden;
 }
 .hashtag {
-  color: white;
-  background-color: #b7b7b7;
+  color: $white;
+  background-color: $gray11;
   border-radius: 22px;
   padding: 0.2rem 0.5rem;
+  &:hover {
+    background-color: $accent-color;
+    transition: all 0.2s ease-in-out;
+  }
 }
