+++ components/AlocholType/ATInfo/ATInfo.module.scss
@@ -1,23 +1,36 @@
 @import '@styles/theme.scss';
 
 .container {
+  width: auto;
+  border-radius: 10px;
+  padding: 0.5rem;
+
   display: flex;
   flex-direction: row;
   gap: 5%;
-  padding: 0.5rem;
-  width: auto;
-  height: 35vh;
-}
-.alcoholImageBox {
-  width: 25%;
-  aspect-ratio: 5/8;
-  border-radius: 10px;
-  box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.1);
+  @media screen and (min-width: $breakpoint-lg) {
+    // desktop
+    width: 100%;
+    max-height: 40vh;
+    overflow-wrap: break-word;
+  }
+  @media screen and (min-width: $breakpoint-sm) and (max-width: $breakpoint-lg) {
+    // tablet
+    height: 25vh;
+  }
+  @media screen and (max-width: $breakpoint-sm) {
+    //mobile
+    height: auto;
+    flex-direction: column;
+  }
 }
 .alcoholImage {
-  width: 100%;
-  height: 100%;
+  width: 25%;
+  min-width: 200px;
+  height: 320px;
   object-fit: cover;
+  border-radius: 10px;
+  box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.1);
 }
 .alcoholInfoBox {
   width: auto;
@@ -25,37 +38,61 @@
   flex-direction: column;
   justify-content: flex-start;
   align-items: left;
-  gap: 0.3rem;
-  font-size: 0.9rem;
+  gap: 0.5rem;
   overflow-x: hidden;
+  @media screen and (max-width: $breakpoint-sm) {
+    gap: 0.7rem;
+  }
 }
 
 .title {
+  max-width: 100%;
   margin: 0;
-  font-size: 2rem;
+
+  font-size: $font-size-xl;
   font-weight: bold;
-  color: #000;
-  margin-bottom: 2rem;
+
+  @include ellipsisText(2, $font-size-xl);
+
+  @media screen and (max-width: $breakpoint-sm) {
+    margin-top: 0.7rem;
+    font-size: $font-size-xl;
+    @include ellipsisText(1, $font-size-xl);
+  }
 }
 .basic {
   margin: 0;
-  font-size: 1rem;
+
+  font-size: $font-size-sm;
+  @include ellipsisText(1, $font-size-md);
+
+  @media screen and (max-width: $breakpoint-sm) {
+    font-size: $font-size-sm;
+    @include ellipsisText(1, $font-size-sm);
+  }
 }
 .hashtagBox {
+  max-height: calc((2 * $font-size-xxs) + 30px);
+
   display: flex;
   flex-direction: row;
   justify-content: flex-start;
   align-items: center;
   flex-wrap: wrap;
-
-  padding-top: 0.5rem;
-
   gap: 10px;
-  font-size: 0.75rem;
+
+  padding-top: 0.2rem;
+
+  font-size: $font-size-xxs;
+  overflow: hidden;
 }
 .hashtag {
-  color: white;
-  background-color: #b7b7b7;
+  color: $white;
+  background-color: $gray11;
   border-radius: 22px;
   padding: 0.2rem 0.5rem;
+  &:hover {
+    background-color: $accent-color;
+    transition: all 0.2s ease-in-out;
+  }
 }
