+++ components/AlocholType/ATInfoDetail/ATInfoDetail.module.scss
@@ -1,3 +1,5 @@
+@import '@styles/theme.scss';
+
 .container {
   display: flex;
   flex-direction: column;
@@ -8,12 +10,12 @@
 }
 .title {
   margin: 0;
-  font-size: 1.8rem;
+  font-size: $font-size-xxl;
   font-weight: bold;
-  color: #000;
   margin-bottom: 1rem;
+  @include ellipsisText(1, $font-size-xxl);
 }
 .basic {
   margin: 0;
-  font-size: 1rem;
+  font-size: $font-size-sm;
 }
