--- components/AlocholType/ATResultList/ATResultList.module.scss
@@ -1,20 +1,0 @@
-.searchList{
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    z-index: 1;
-    border-radius: 20px;
-    text-align: left;
-    background-color: #fff;
-}
-.searchList > li{
-    width: 80%;
-    background-color: #fff;
-    padding: 10px 0 10px 1.3rem;
-}
-.searchList > li:hover{
-    background-color: #f5f5f5;
-}
