+++ components/AlocholType/ATSearchBar/ATSearchBar.module.scss
@@ -1,38 +1,89 @@
-.searchBarBox{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    align-self: center;
-    padding: 5rem 0;
-    height: 50px;
-    border-radius: 20px;
-    text-align: center;
-}
-.searchBar{
-    display: flex;
-    align-items: center;
-    width: 80%;
-    box-shadow: rgba(0,0,0,0.3) 1px 1px 1px 1px;
-    border-radius: 20px;
-    border: rgba(0,0,0,0.3) 1px solid;
-}
-.searchInput{
-    outline: rgba(0,0,0,0.5) 1px solid;
-    background-color:white;
-    color: #000;
-    width: 90%;
-    height: 3rem;
-    border-radius: 20px 0 0 20px;
+@import '@styles/theme.scss';
 
-    font-size: 1.2rem;
-    font-weight: regular;
-    text-align: left;
-    padding-left: 1.3rem;
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  padding: 0.5rem;
+  width: auto;
+  margin: 1.5rem 0 2rem 0;
 }
-.searchButton{
-    width: 10%;
-    height: 100%;
-    background-color: #000;
-    border-radius: 0 20px 20px 0;
+.title {
+  margin: 0;
+  font-size: $font-size-xxl;
+  font-weight: bold;
+  margin-bottom: 1rem;
+  @include ellipsisText(1, $font-size-xxl);
+}
+.basic {
+  margin: 0;
+  font-size: $font-size-sm;
+}
+
+.searchBarBox {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  align-self: center;
+  padding: 5rem 0;
+  height: 50px;
+  border-radius: 20px;
+  text-align: center;
+}
+.searchBar {
+  display: flex;
+  align-items: center;
+  width: 80%;
+  border-radius: 20px;
+  border: 0;
+  border-bottom: 2px solid $gray11;
+}
+.searchInput {
+  background-color: $white;
+  width: 90%;
+  height: 3rem;
+  border: 0;
+  border-radius: 20px 0 0 20px;
+  outline: none;
+
+  font-size: $font-size-md;
+  font-weight: regular;
+  text-align: left;
+  padding-left: 1.3rem;
+  background-color: rgba($color: $gray11, $alpha: 0.04);
+  &:focus {
+    background-color: rgba($color: $gray11, $alpha: 0.15);
+  }
+}
+.searchButton {
+  width: 10%;
+  height: 100%;
+  border-radius: 0 20px 20px 0;
+  border: 0 1px 2px 0 solid $gray11;
+
+  background: rgba($accent-color, 0.5);
+}
+
+.searchList {
+  width: 100%;
+  height: 100%;
+  border-radius: 10px;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  text-align: left;
+
+  background-color: $white;
+}
+.searchList > li {
+  width: 80%;
+  background-color: $white;
+  padding: 10px 0 10px 1.3rem;
+  border-radius: 10px;
+}
+.searchList > li:hover {
+  background-color: rgba($color: $gray11, $alpha: 0.3);
 }
